pip_requirements.txt
flask
Pillow
numpy
